pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-scheme_env"
version = "0.1.0"
description = "Indian Government Welfare Officer RL Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "openai>=2.7.2,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "scheme_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["scheme_env", "scheme_env.server"]
package-dir = { "scheme_env" = ".", "scheme_env.server" = "server" }